requirements.txt
# --- Core web framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# --- Document parsing ---
pymupdf>=1.23.7      # PDF extraction
python-docx>=0.8.11  # DOCX extraction

# --- Chunking / LLM pipeline ---
langchain>=0.1.15

# --- Embedding & Vector search ---
sentence-transformers>=2.6.1
faiss-cpu>=1.7.4
# Optional alternative: uncomment if you prefer Chroma
# chromadb>=0.4.24

# --- Local LLM runtime ---
llama-cpp-python>=0.2.45
# Optional alternative: uncomment if using GPT4All
# gpt4all>=2.1.2

# --- Front‑end UI ---
streamlit>=1.33.0

# --- ORM / Database layer ---
sqlalchemy>=2.0.29

# --- Utilities ---
requests>=2.31.0
pandas>=2.2.2
python-dotenv>=1.0.1
tqdm>=4.66.2
python-multipart>=0.0.6
